Schulwegsicherheit: Move shared LegendStyles into Component
Use by ZESPlus only, until now. And now shared between pages.

Delete Legend.tsx: File was empty.
